+++ CHANGELOG.md
@@ -2,11 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## v2.0.0 - 2018-11-12
+
+### Changes
+
+- Rename `encrypt.sync` and `decrypt.sync` to `encrypt.aead` and `decrypt.aead`
+- Rename `encrypt.async` and `decrypt.async` to `encrypt.pki` and `decrypt.pki`
+
 ## v1.0.3 - 2018-09-21
 
 ### Misc
 
-- Change repo name to @auth0/magic
+- Change repo name to auth0-magic
 
 ## v1.0.1 - 2018-09-14